+++ Sources/Resources/StoryboardResource.swift
@@ -9,11 +9,35 @@
 /// A protocol that describes an item that can represent a storyboard.
 public protocol StoryboardDescribing {
 
-    /// The `String` that will be used to represent the name of the storyboard.
+    /// The `String` that represents the name of the storyboard.
     var name: String { get }
 
-    /// The `Bundle` that will be used to represent the bundle of the storyboard.
+    /// The `Bundle` that represents the bundle of the storyboard.
     var bundle: Bundle { get }
+
+}
+
+public extension StoryboardDescribing {
+
+    var bundle: Bundle {
+        return Bundle.main
+    }
+
+}
+
+/// A struct that describes an item that can represent a storyboard.
+public struct StoryboardDescription: StoryboardDescribing {
+
+    /// The `String` that represents the name of the storyboard.
+    public let name: String
+
+    /// The `Bundle` that represents the bundle of the storyboard.
+    public let bundle: Bundle
+
+    public init(name: String, bundle: Bundle = Bundle.main) {
+        self.name = name
+        self.bundle = bundle
+    }
 
 }