+++ Tests/ResourcesTests/AlertActionResourceTests.swift
@@ -1,6 +1,6 @@
 //
 //  AlertActionResourceTests.swift
-//  Resource
+//  Tests
 //
 //  Created by Justin Jia on 12/5/16.
 //  Copyright © 2016 TintPoint. MIT license.
@@ -9,7 +9,7 @@
 import XCTest
 @testable import Resource
 
-enum AppAlertAction: String {
+enum AppAlertAction {
 
     case test
 
@@ -17,8 +17,8 @@ enum AppAlertAction: String {
 
 extension AppAlertAction: AlertActionDescribing {
 
-    var title: String {
-        return rawValue.capitalized
+    var title: String? {
+        return "Test"
     }
 
     var style: UIAlertActionStyle {
@@ -33,10 +33,16 @@ extension AppAlertAction: AlertActionDescribing {
 
 class AlertActionResourceTests: XCTestCase {
 
-    func testAlertActionResource() {
+    func testAlertActionProtocolResource() {
         let action = Resource.of(AppAlertAction.test)
         XCTAssertEqual(action.title, "Test")
         XCTAssertEqual(action.style, .default)
     }
-    
+
+    func testAlertActionStructResource() {
+        let action = Resource.of(AlertActionDescription(title: "Title", style: .cancel, handler: nil))
+        XCTAssertEqual(action.title, "Title")
+        XCTAssertEqual(action.style, .cancel)
+    }
+
 }