+++ Tests/ResourcesTests/LocalizedStringResourceTests.swift
@@ -1,26 +1,32 @@
 //
-//  LocalizedUserNotificationStringResourceTests.swift
-//  Resource
+//  LocalizedStringResourceTests.swift
+//  Tests
 //
-//  Created by Justin Jia on 11/14/16.
+//  Created by Justin Jia on 11/6/16.
 //  Copyright © 2016 TintPoint. MIT license.
 //
 
 import XCTest
 @testable import Resource
 
-enum AppLocalizedUserNotificationString {
+enum AppLocalizedString {
 
     case test(argument: String?)
 
 }
 
-extension AppLocalizedUserNotificationString: LocalizedStringDescribing {
+extension AppLocalizedString: LocalizedStringDescribing {
 
     var key: String {
         switch self {
         case .test(let argument) where argument != nil: return "With Argument %@"
         case .test: return "No Argument"
+        }
+    }
+
+    var comments: String {
+        switch self {
+        case .test: return "Comment"
         }
     }
 
@@ -33,11 +39,16 @@ extension AppLocalizedUserNotificationString: LocalizedStringDescribing {
 
 }
 
-class LocalizedUserNotificationStringResourceTests: XCTestCase {
+class LocalizedStringResourceTests: XCTestCase {
 
-    func testLocalizedUserNotificationStringResource() {
+    func testLocalizedStringProtocolResource() {
         XCTAssertEqual(Resource.of(AppLocalizedString.test(argument: "A")), "With Argument A")
         XCTAssertEqual(Resource.of(AppLocalizedString.test(argument: nil)), "No Argument")
     }
-    
+
+    func testLocalizedStringStructResource() {
+        XCTAssertEqual(Resource.of(LocalizedStringDescription(key: "Key %@", comment: "Comment", arguments: ["X"])), "Key X")
+        XCTAssertEqual(Resource.of(LocalizedStringDescription(key: "Key", comment: "Comment", arguments: nil)), "Key")
+    }
+
 }