+++ Tests/ResourcesTests/StringResourceTests.swift
@@ -1,6 +1,6 @@
 //
 //  StringResourceTests.swift
-//  Resource
+//  Tests
 //
 //  Created by Justin Jia on 11/6/16.
 //  Copyright © 2016 TintPoint. MIT license.
@@ -9,7 +9,7 @@
 import XCTest
 @testable import Resource
 
-enum AppString: String {
+enum AppString {
 
     case test
 
@@ -18,15 +18,19 @@ enum AppString: String {
 extension AppString: StringDescribing {
 
     var content: String {
-        return rawValue
+        return "Test"
     }
 
 }
 
 class StringResourceTests: XCTestCase {
 
-    func testStringResource() {
-        XCTAssertEqual(Resource.of(AppString.test), AppString.test.rawValue)
+    func testStringProtocolResource() {
+        XCTAssertEqual(Resource.of(AppString.test), "Test")
+    }
+
+    func testStringStructResource() {
+        XCTAssertEqual(Resource.of(StringDescription(content: "Content")), "Content")
     }
 
 }